requirements.txt
anchor-exp>=0.0.2.0,<1
dalex>=1.5.0,<2
dice-ml==0.11
dill>=0.3.9,<4
Flask[async]
flask-cors>=3.0.13
gin-config==0.5.0
gunicorn==23.0.0
lime==0.2.0.1
llama-index==0.12.23
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.23.post2
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.25
llama-index-program-openai==0.3.1
llama-index-utils-workflow==0.3.0
matplotlib==3.7.0
numpy>=1.23.5,<2.0.0
openai==1.62.0
agents>=0.1.0
pandas==1.5.3
pandas-stubs==1.2.0.53
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0.2
regex==2022.3.15
requests==2.32.3
s3transfer==0.5.2
sacremoses==0.0.49
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers
sentencepiece
sentry-sdk==1.9.5
setproctitle==1.3.2
shap==0.40.0
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
tifffile==2022.3.25
tokenizers==0.11.6
tomli==2.0.1
tqdm==4.66.5
transformers==4.17.0
twilio==7.8.0
typing_extensions==4.12.2
urllib3==1.26.12
wandb==0.13.2
Werkzeug==2.0.3
word2number==1.1
wordninja==2.0.0
openai
anchor-exp>=0.0.2.0,<1
dalex>=1.5.0,<2
seaborn==0.12.2
kaleido
huggingface-hub==0.4.0
python-dotenv==1.0.1
tenacity==8.3.0
transitions>=0.9.1,<1
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn
scipy
shap
sklearn
tqdm==4.66.5
typing_extensions
transitions
seaborn
setuptools
PyPDF2>=3.0.0
